backend/requirements.txt
flask==2.3.3 
flask-cors==4.0.0 
python-dotenv==1.0.0 
requests==2.31.0 
python-jose==3.3.0 
sqlalchemy==2.0.23 
flask-sqlalchemy==3.1.1 
flask-migrate==4.0.5 
